Extract hypervisor process identity logic into process_identity.go
